% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{concweb}
\alias{concweb}
\title{Concurrency Simulation Shiny App}
\usage{
concweb()
}
\description{
Runs a web browser-based GUI of the concurrency microsimulation.
}
\examples{
\dontrun{
concweb()
}

}
\keyword{GUI}
