% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{transco}
\alias{transco}
\title{Obtain Transparent Colors}
\usage{
transco(col, alpha = 1, invisible = FALSE)
}
\arguments{
\item{col}{Vector consisting of colors, of any of the three kinds of
\code{R} color specifications (named, hexadecimal, or integer; see
\code{\link{col2rgb}}).}

\item{alpha}{Vector of transparency levels, where 0 is transparent and 1
is opaque.}

\item{invisible}{Supresses printing of the RGB color.}
}
\value{
A vector of length equal to the input \code{col} vector or the \code{alpha},
vector, if one or the other is of length greater than 1, containing the
transformed color values in hexidemical format.
}
\description{
Returns an RGB transparent color from any standard R color.
}
\details{
The purpose of this function is to facilitate color transparency, which is
used widely in \code{EpiModel} plots. This is an internal function that is
not ordinarily called by the end-user. This function allows that one of col
or alpha may be of length greater than 1.
}
\seealso{
\code{\link{rgb}}, \code{\link{col2rgb}}
}
\keyword{internal}

