% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrency.sim-package.R
\docType{package}
\name{concurrency.sim-package}
\alias{concurrency.sim-package}
\title{concurrency.sim Package}
\description{
\tabular{ll}{
   Package: \tab concurrency.sim\cr
   Type: \tab Package\cr
   Version: \tab 0.6\cr
   Date: \tab 2018-04-03\cr
   License: \tab GPL (>= 2)\cr
   LazyLoad: \tab yes\cr
}
}
\details{
Microsimulation model for simulating the effects of sexual partnership
concurrency on the HIV-1 transmission dynamics in a heterosexual epidemic
context.
}
\keyword{package}
